requirements.txt
fastapi==0.95.2
pydantic==1.10.9
strawberry-graphql
uvicorn~=0.18.2
sqlalchemy~=2.0.35
alembic~=1.13.3
h11~=0.13.0
pip~=24.0
wheel~=0.43.0
Mako~=1.3.5
Pygments~=2.13.0
MarkupSafe~=3.0.1
anyio~=3.6.1
sniffio~=1.2.0
setuptools~=69.2.0
python-dateutil~=2.9.0.post0
starlette~=0.27.0
requests~=2.28.1
click~=8.1.3
asgiref~=3.6.0
six~=1.15.0
greenlet~=3.1.1
decorator~=5.1.1
Django~=4.1.4
psycopg2
isort
pika
kombu